podcast-summarizer/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
feedparser==6.0.10
openai==1.109.1
sqlalchemy==2.0.23
alembic==1.13.1
httpx==0.28.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.10.6
pydantic-settings==2.1.0
youtube-transcript-api==1.2.2
python-dateutil==2.8.2
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.185.0
lxml==6.0.2
assemblyai==0.45.4
exa_py==1.14.20
langgraph>=0.6.11
langchain>=0.3.27
langchain-openai>=0.3.35
langsmith>=0.4.37
psycopg2-binary>=2.9.9

